requirements.txt
black==22.6.0
Flask==2.2.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
marshmallow==3.17.0
opencv-python==4.6.0.66
pandas==1.4.3
seaborn==0.11.2
Pillow==9.2.0
PyYAML==6.0
requests==2.28.1
scipy==1.9.0
torch==1.9.1
torchvision==0.10.1
tqdm==4.64.0
